+++ src/root/views/TranslationDashboard/index.js
@@ -3,6 +3,9 @@ import ContentEditable from 'react-contenteditable';
 import { connect } from 'react-redux';
 import _cs from 'classnames';
 import spark from 'spark-md5';
+import Helmet from 'react-helmet';
+import sheet from 'xlsx';
+
 import {
   listToMap,
   isDefined,
@@ -21,17 +24,21 @@ import {
   languageResponseSelector,
 } from '#selectors';
 
+
 import BlockLoading from '#components/block-loading';
 import LanguageSelect from '#components/LanguageSelect';
 
 import styles from './styles.module.scss';
 
 // TODO: use translation
-const views = {
-  all: 'All in server',
+const conflictedViews = {
   added: 'New by dev',
   removed: 'Removed by dev',
   updated: 'Updated by dev',
+};
+
+const modifiableViews = {
+  all: 'All in server',
 };
 
 function StringRow(p) {
@@ -70,6 +77,7 @@ function StringRow(p) {
   );
 }
 
+
 function TranslationDashboard(p) {
   const {
     className,
@@ -81,8 +89,20 @@ function TranslationDashboard(p) {
     getLanguage,
   } = p;
 
-  const viewKeys = React.useMemo(() => Object.keys(views), []);
-  const [currentView, setCurrentView] = React.useState(viewKeys[0]);
+  const handleExportButtonClick = React.useCallback(() => {
+    const langEntries = Object.entries(lang);
+    const ws = sheet.utils.aoa_to_sheet([
+      ['ID', 'dev', 'en', 'fr', 'es', 'ar'],
+      ...langEntries,
+    ]);
+    const wb = sheet.utils.book_new();
+    sheet.utils.book_append_sheet(wb, ws);
+
+    sheet.writeFile(wb, 'go-strings.xlsx');
+  }, []);
+
+  const conflictedViewKeys = React.useMemo(() => Object.keys(conflictedViews), []);
+  const [currentView, setCurrentView] = React.useState(conflictedViewKeys[0]);
   const prevBulkResponse = React.useRef(languageBulkResponse);
 
   React.useEffect(() => {
@@ -94,16 +114,20 @@ function TranslationDashboard(p) {
     prevBulkResponse.current = languageBulkResponse;
   }, [prevBulkResponse, languageBulkResponse, currentLanguage, getLanguage]);
 
-  const appStrings = React.useMemo(() => (
-    listToMap(
-      languageData.strings || [],
-      d => d.key,
-      d => ({
-        hash: d.hash,
-        value: d.value
-      }),
-    )
-  ), [languageData]);
+  const [appStrings, setAppStrings] = React.useState({});
+
+  React.useEffect(() => {
+    setAppStrings(
+      listToMap(
+        languageData.strings || [],
+        d => d.key,
+        d => ({
+          hash: d.hash,
+          value: d.value
+        }),
+      )
+    );
+  }, [setAppStrings, languageData]);
 
   const appStringKeyList = React.useMemo(() => {
     const keys = Object.keys(appStrings);
@@ -134,9 +158,20 @@ function TranslationDashboard(p) {
       ...oldStrings,
       [key]: {
         value,
+        hash: oldStrings[key].hash,
       }
     }));
   }, [setStrings]);
+
+  const handleAppStringChange = React.useCallback((key, value) => {
+    setAppStrings((oldStrings) => ({
+      ...oldStrings,
+      [key]: {
+        value,
+        hash: oldStrings[key].hash,
+      }
+    }));
+  }, [setAppStrings]);
 
   const [addedKeyList, removedKeyList, updatedKeyList] = React.useMemo(() => {
     const allKeyList = [...new Set([...devStringKeyList, ...appStringKeyList])];
@@ -172,16 +207,16 @@ function TranslationDashboard(p) {
   const removedKeys = React.useMemo(() => listToMap(removedKeyList, d => d, d => true), [removedKeyList]);
 
   const handleSaveButtonClick = React.useCallback(() => {
-    const actions = Object.keys(strings).map((key) => ({
+    const actions = Object.keys(appStrings).map((key) => ({
       action: 'set',
       key,
-      value: strings[key].value,
-      hash: strings[key].hash,
+      value: appStrings[key].value,
+      hash: appStrings[key].hash,
     }));
 
     const data = { actions };
     postLanguageBulk(currentLanguage, data);
-  }, [strings, postLanguageBulk, currentLanguage]);
+  }, [appStrings, postLanguageBulk, currentLanguage]);
 
   const handleRemoveOutdatedButtonClick = React.useCallback(() => {
     const actions = removedKeyList.map((key) => ({
@@ -192,6 +227,35 @@ function TranslationDashboard(p) {
     const data = { actions };
     postLanguageBulk(currentLanguage, data);
   }, [removedKeyList, postLanguageBulk, currentLanguage]);
+
+  const handleResolveButtonClick = React.useCallback(() => {
+    const actions = updatedKeyList.map((key) => {
+      const value = strings[key].value;
+      const devValue = devStrings[key].value;
+
+      return {
+        action: 'set',
+        key,
+        value,
+        hash: spark.hash(devValue),
+      };
+    });
+
+    const data = { actions };
+    postLanguageBulk(currentLanguage, data);
+  }, [strings, updatedKeyList, devStrings, postLanguageBulk, currentLanguage]);
+
+  const handleAddNewKeysButtonClick = React.useCallback(() => {
+    const actions = addedKeyList.map((key) => ({
+      action: 'set',
+      key,
+      value: devStrings[key].value,
+      hash: devStrings[key].hash,
+    }));
+
+    const data = { actions };
+    postLanguageBulk(currentLanguage, data);
+  }, [devStrings, addedKeyList, postLanguageBulk, currentLanguage]);
 
   const handleTabClick = React.useCallback((e) => {
     setCurrentView(e.target.name);
@@ -215,27 +279,76 @@ function TranslationDashboard(p) {
   }), [appStringKeyList, addedKeyList, removedKeyList, updatedKeyList]);
 
   const pending = languageResponse.fetching || languageBulkResponse.fetching;
+  const conflicted = React.useMemo(() => (
+    viewCounts.added > 0 || viewCounts.removed > 0 || viewCounts.updated > 0
+  ), [viewCounts]);
+  const views = conflicted ? conflictedViews : modifiableViews;
+
+  const handleFileInputChange = React.useCallback((e) => {
+    const file = e.target.files[0];
+    if (file) {
+      const reader = new FileReader();
+      reader.onload = (e) => {
+        const workbook = sheet.read(e.target.result, { type: 'binary' });
+        const firstSheet = workbook.SheetNames[0];
+        const rowList = sheet.utils.sheet_to_row_object_array(workbook.Sheets[firstSheet]);
+        const rows = listToMap(rowList, d => d['ID'], d => d[currentLanguage]);
+        const newAppStrings = devStringKeyList.reduce((acc, key) => {
+          acc[key] = {
+            value: rows[key],
+            hash: devStrings[key].hash,
+          };
+          return acc;
+        }, {});
+
+        setAppStrings(newAppStrings);
+      };
+      reader.readAsBinaryString(file);
+    }
+  }, [devStringKeyList, devStrings, setAppStrings, currentLanguage]);
 
   return (
     <div className={_cs(className, styles.translationDashboard)}>
+      <Helmet>
+        <title>
+          IFRC GO - Translation Dashboard
+        </title>
+      </Helmet>
       <header className={styles.header}>
         <div className={styles.topSection}>
           <h2 className={styles.heading}>
-            Translation
+            Translation Dashboard
           </h2>
           <div className={styles.actions}>
+            <label htmlFor="import" className={_cs(pending && 'disabled', 'button button--secondary-bounded')}>
+              Import from xlsx
+            </label>
+            <input
+              disabled={pending}
+              id="import"
+              type="file"
+              accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+              onChange={handleFileInputChange}
+              hidden
+            />
+            <button
+              onClick={handleExportButtonClick}
+              className='button button--secondary-bounded'
+            >
+              Export dev strings
+            </button>
             <LanguageSelect />
           </div>
         </div>
         <div className={styles.bottomSection}>
           <div className={styles.tabs}>
-            { viewKeys.map(viewKey => (
+            { Object.keys(views).map(viewKey => (
               <button
                 key={viewKey}
                 name={viewKey}
                 onClick={handleTabClick}
                 type="button"
-                className={_cs(styles.tab, currentView === viewKey && styles.active)}
+                className={_cs(styles.tab, (!conflicted || currentView === viewKey) && styles.active)}
                 disabled={pending}
               >
                 {`${views[viewKey]} (${viewCounts[viewKey]})`}
@@ -243,17 +356,39 @@ function TranslationDashboard(p) {
             ))}
           </div>
           <div className={styles.tabActions}>
-            { currentView === 'removed' ? (
-              <button
-                className="button button--secondary-bounded"
-                onClick={handleRemoveOutdatedButtonClick}
-                disabled={pending}
-              >
-                Remove outdated keys
-              </button>
+            { conflicted ? (
+              <>
+                { currentView === 'removed' && (
+                  <button
+                    className={_cs(pending && 'disabled', 'button button--secondary-bounded')}
+                    onClick={handleRemoveOutdatedButtonClick}
+                    disabled={pending}
+                  >
+                    Remove outdated keys
+                  </button>
+                )}
+                { currentView === 'added' && (
+                  <button
+                    className={_cs(pending && 'disabled', 'button button--secondary-bounded')}
+                    onClick={handleAddNewKeysButtonClick}
+                    disabled={pending}
+                  >
+                    Add new keys
+                  </button>
+                )}
+                { currentView === 'updated' && (
+                  <button
+                    className={_cs(pending && 'disabled', 'button button--secondary-bounded')}
+                    onClick={handleResolveButtonClick}
+                    disabled={pending}
+                  >
+                    Resolve
+                  </button>
+                )}
+              </>
             ) : (
               <button
-                className="button button--primary-bounded"
+                className={_cs(pending && 'disabled', 'button button--primary-bounded')}
                 onClick={handleSaveButtonClick}
                 disabled={pending}
               >
@@ -280,49 +415,49 @@ function TranslationDashboard(p) {
           { pending ? (
             <BlockLoading />
           ) : (
-            <>
-              { currentView === 'all' && appStringKeyList.map((k) => (
-                <StringRow
-                  key={k}
-                  stringKey={k}
-                  devValue={devStrings[k]?.value}
-                  value={strings[k]?.value || appStrings[k]?.value}
-                  editable={!removedKeys[k]}
-                  obsolete={removedKeys[k]}
-                  onChange={handleStringChange}
-                />
-              ))}
-              { currentView === 'added' && addedKeyList.map((k) => (
-                <StringRow
-                  key={k}
-                  stringKey={k}
-                  devValue={devStrings[k]?.value}
-                  value={strings[k]?.value || appStrings[k]?.value}
-                  editable={!removedKeys[k]}
-                  obsolete={removedKeys[k]}
-                  onChange={handleStringChange}
-                />
-              ))}
-              { currentView === 'removed' && removedKeyList.map((k) => (
+            conflicted ? (
+              <>
+                { currentView === 'added' && addedKeyList.map((k) => (
+                  <StringRow
+                    key={k}
+                    stringKey={k}
+                    devValue={devStrings[k]?.value}
+                    value={devStrings[k]?.value}
+                  />
+                ))}
+                { currentView === 'removed' && removedKeyList.map((k) => (
+                  <StringRow
+                    key={k}
+                    stringKey={k}
+                    devValue={devStrings[k]?.value}
+                    value={appStrings[k]?.value}
+                  />
+                ))}
+                { currentView === 'updated' && updatedKeyList.map((k) => (
+                  <StringRow
+                    key={k}
+                    stringKey={k}
+                    devValue={devStrings[k]?.value}
+                    value={strings[k]?.value}
+                    editable={!removedKeys[k]}
+                    obsolete={removedKeys[k]}
+                    onChange={handleStringChange}
+                  />
+                ))}
+              </>
+            ) : (
+              appStringKeyList.map((k) => (
                 <StringRow
                   key={k}
                   stringKey={k}
                   devValue={devStrings[k]?.value}
                   value={appStrings[k]?.value}
-                />
-              ))}
-              { currentView === 'updated' && updatedKeyList.map((k) => (
-                <StringRow
-                  key={k}
-                  stringKey={k}
-                  devValue={devStrings[k]?.value}
-                  value={strings[k]?.value || appStrings[k]?.value}
                   editable={!removedKeys[k]}
                   obsolete={removedKeys[k]}
-                  onChange={handleStringChange}
+                  onChange={handleAppStringChange}
                 />
-              ))}
-            </>
+              ))
+            )
           )}
         </div>
       </div>