+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Release 4.3.12
+
+Hotfix release: Fix undefined string for Potentially Affected in Field Report frontend
+
 ### Release 4.3.11
 
 This release includes a lot of code changes to support future translations support, as well as significant changes toward implementing updated site designs.
@@ -13,6 +17,7 @@ Frontend:
    - Styling cleanup of Highlighted Operations + Adding Follow buttons: https://github.com/IFRCGo/go-frontend/issues/1206
  - Improvements to Signup Workflow: https://github.com/IFRCGo/go-frontend/issues/1233
  - Fix Create a Report dropdown on mobile: https://github.com/IFRCGo/go-frontend/issues/1363
+ - Rename Subscribe / Unsubscribe to Follow / Unfollow on Emergency page: https://github.com/IFRCGo/go-frontend/issues/1370
 
 Backend: